Fix [igzValidatingInputField] Cannot share validation rules
If the same array is passed to `validationRules` attribute of different
instances of `igzValidatingInputField` - it breaks.
One of the instances sets rules valid/invalid, and it is reflected in
the other instance, too.

In addition, changed `validationRules` attribute from `<` to `<*` so
added/removed validation rules are detected without changing to a
different array.